requirements.txt
flask==2.0.1
Pillow==10.2.0
pytesseract==0.3.8
python-dotenv==0.19.0
Werkzeug==2.0.1 